requirements.txt
Flask
flask-compress
Jinja2
MarkupSafe
Werkzeug
certifi
chardet
click
configparser
gunicorn
idna
itsdangerous
lxml
mysql-connector
pip
pylint
requests
setuptools
urllib3
wheel